[IMP] hr_holidays: Improve leave type: limit date, visibility, usability
Purpose
=======

Caroline needs help to limit leaves recording errors. We need to add some
configuration on leave type object. We have to avoid poeple taking legal
leaves 2016 on 2017.

Specification
=============

Add validity dates: start and end dates on the leave type.

Double validation: change the checkbox into a radio button

Add a sequence field

Add a field, to specify if the employee can or not select it in his
request.