[IMP] Accrual leaves support added
Purpose
=======

We need to support the accrual allocation for leaves. For example in the
USA they need to allocate 2 hours of leave every 2 weeks of work.

Specification
=============

Add the possibility to create accrual allocations

Add a default start date on the day accrual allocations are created

Add a field on hr.leave.type so that it is selectable as an accrual
leave type

If, in the period of the accrual, some "unpaid" leaves has been taken,
compute the accrual in the prorata.