[IMP] base,web: views: 'invalidate_cache' attribute
This rev. adds the support of the 'invalidate_cache' attribute in
views, e.g.

`<tree editable="top" invalidate_cache="1">...</tree>`

When set to "1" on the root node of an arch, the  data_manager's
cache is invalidated at each create, update and unlink operation.
This cache stores actions (results of calls to /web/action/load
route), views (results of load_views RPCs) and filters (fetched
alonside views).

This new attribute is thus especially useful for views allowing
to create, edit or delete records of those models, as it spares
the user from reloading the webclient after performing such
operations to see its changes in the interface. With the attribute
set, he only has to replay the action to see its changes.