[IMP] hr_holidays: display number of hours when in hourly mode
Purpose
=======

When taking the leaves in hours, it is better to show the number of
hours taken instead of showing the number of days taken.

Specification
=============

Internally we keep a count of the number of days taken, but we convert
it in hours in order to show it to the user.