[FIX] web: kanban: action button check the domain to display the record
When a user click on an action button, we need to check if the record match
with the domain to remove it from the view if this action change the record
value.
Special case: we filter by active field equal true, if the active field is
available in the data (field present in the kanban view).