[IMP] chat, hr: add a mixin to manage partner creation / subscription on record creation by email or email answer.
This mixin aims to replace some redundant logic in some applications and to
autosubscribe a partner when replying on a record by email.
The main problem is linked to the fact that when a record is created by email
the partner is not added to the followers. He will never get the answer. Adding
this mixin will solve the problem by creating a partner if it doesn't exists and
subscibing it when needed. Moreover redundant logic, like linking the new partner
to a specific field of the record have been added to this mixin.
The hr_recruitment app have been updated to use this mixin for test and demonstration purpose.