[ADD] ir_ui_view: restore views to default
With this commit, if a user modifies a view it is now possible to
restore it to its original state by reloading the content from the XML
file that defines the view.

The restore can be activated either via RPC by calling the restore_views
method on a ir.ui.view recordset, or from the UI in the form view of
ir.ui.view records (debug mode must be activated).

One of the seven deadly sins.