[IMP] resource/resource_mixin : Add possibility to pass in a domain
Purpose
=======

It will enable to retrieve the number of work and leave days of an
employee according to a certain domain. Which can be used in order to
filter on certain leave types only.

Specification
=============

If, in the period of the accrual, some "unpaid" leaves has been taken,
compute the accrual in the prorata.

In resource_mixin and resource add a way to use a domain in order to
filter the leaves.